Ported TestDateTimeFormatter to scalatest
